task class refactoring
