+++ experiment_scheduler/task_manager/task_manager_server.py
@@ -3,12 +3,11 @@
 import os
 import subprocess
 import ast
-from typing import Dict, List, Optional, Union
+from typing import Dict, Optional
 from concurrent import futures
 from os import path as osp
 
 import grpc
-import psutil
 import pynvml
 
 from experiment_scheduler.common.logging import get_logger, start_end_logger
@@ -23,81 +22,12 @@ from experiment_scheduler.task_manager.grpc_task_manager.task_manager_pb2 import
     ProgressResponse,
 )
 from experiment_scheduler.task_manager.return_code import ReturnCode
+from experiment_scheduler.task_manager.task import Task
 
 logger = get_logger(name="task_manager")
 
 KILL_CHILD_MAX_DEPTH = 2
 CHUNK_SIZE = 1024 * 5
-
-
-class ProcessUtil:
-    """Wrapper class for psutil.
-
-    Args:
-        pid (int): process id.
-    """
-
-    def __init__(self, pid: int):
-        self._pid = pid
-        try:
-            self._process = psutil.Process(self._pid)
-        except psutil.NoSuchProcess:
-            self._process = None
-        except psutil.AccessDenied:
-            logger.warning("Access to process(%d) is denied.", pid)
-            self._process = None
-
-    @property
-    def pid(self):
-        """
-        return process id
-        :return: pid
-        """
-        return self._pid
-
-    def kill_itself_with_child_process(self, max_depth: int) -> bool:
-        """kill self recursively up to max depth.
-
-        Args:
-            max_depth (int): max depth of child processes to find and kill.
-
-        Returns:
-            bool:
-                Whether success to kill process recursively.
-        """
-        if self._process is None:
-            logger.warning("Fail to get process(%d).", self._pid)
-            return False
-        return self.kill_process_recursively(self._process, max_depth, 0)
-
-    @staticmethod
-    def kill_process_recursively(
-        process: psutil.Process, max_depth: int, cur_depth: int = 0
-    ) -> bool:
-        """kill process recursively up to max depth.
-
-        Args:
-            max_depth (int): max depth of child processes to find and kill.
-            cur_depth (int): current depth.
-
-        Returns:
-            bool:
-                Whether success to kill process recursively.
-        """
-        if not process.is_running():
-            return True
-
-        if cur_depth < max_depth:
-            for child_process in process.children():
-                ProcessUtil.kill_process_recursively(
-                    child_process, max_depth, cur_depth + 1
-                )
-
-        if process.is_running():
-            process.terminate()
-            process.wait()
-
-        return True
 
 
 class ResourceManager:
@@ -176,32 +106,6 @@ class ResourceManager:
         return list(self._resource_rental_history.keys())
 
 
-class Task:
-    """Wrapper class for Task."""
-
-    def __init__(self, process: subprocess.Popen):
-        self._process = process
-        self._history: List[Dict[str, float]] = []
-
-    def get_return_code(self) -> Optional[int]:
-        """Get return code."""
-        return self._process.poll()
-
-    def register_progress(self, progress: Union[int, float], leap_second: float):
-        """Register progress."""
-        self._history.append({"progress": progress, "leap_second": leap_second})
-
-    def get_progress(self) -> Optional[Union[int, float]]:
-        """Get latest progress."""
-        if self._history:
-            return self._history[-1]["progress"]
-        return None
-
-    def get_pid(self) -> int:
-        """Get pid."""
-        return self._process.pid
-
-
 class TaskManagerServicer(task_manager_pb2_grpc.TaskManagerServicer, ReturnCode):
     """Provides methods that implement functionality of task manager server."""
 
@@ -260,7 +164,7 @@ class TaskManagerServicer(task_manager_pb2_grpc.TaskManagerServicer, ReturnCode)
             ),
             stderr=subprocess.STDOUT,
         )
-        self.tasks[task_id] = Task(task)
+        self.tasks[task_id] = Task(task.pid)
 
         self.logger.info("%s is now running!", task_id)
 
@@ -320,8 +224,7 @@ class TaskManagerServicer(task_manager_pb2_grpc.TaskManagerServicer, ReturnCode)
 
         if sign is not None:
             return TaskStatus(task_id=request.task_id, status=TaskStatus.Status.DONE)
-        p_util = ProcessUtil(target_process.pid)
-        if p_util.kill_itself_with_child_process(KILL_CHILD_MAX_DEPTH):
+        if target_process.kill_itself_with_child_process(KILL_CHILD_MAX_DEPTH):
             self.logger.info("%s is killed!", request.task_id)
             return TaskStatus(task_id=request.task_id, status=TaskStatus.Status.KILLED)
         self.logger.info("Fail to kill %s", request.task_id)
@@ -383,20 +286,9 @@ class TaskManagerServicer(task_manager_pb2_grpc.TaskManagerServicer, ReturnCode)
 
     def _find_which_task_report(self, current_pid: int) -> Optional[Task]:
         """find out all pids of task considering a case that pid is child process of task"""
-        tasks_pid = {task.get_pid(): task for task in self.tasks.values()}
-        pid_hierachy = [current_pid]
-
-        try:
-            current_process = psutil.Process(current_pid)
-        except psutil.NoSuchProcess:
-            return None
-
-        pid_hierachy += [process.pid for process in current_process.parents()]
-
-        for pid in pid_hierachy:
-            if pid in tasks_pid:
-                return tasks_pid[pid]
-
+        for task in self.tasks.values():
+            if current_pid == task.pid or task.is_child_pid(current_pid):
+                return task
         return None
